Trying to make cub to sphere 173
Last time when I made the dimensions of the sphere 5,32,32 it did not work but 1,6,6 created an odd shaped sphere. 5,32,32 is how to make a perfect sphere but its not working and im assuming that is because it is too big. Im trying to make it smaller but still spherical.